+++ airflow/dag_processing/manager.py
@@ -661,10 +661,7 @@ class DagFileProcessorManager(LoggingMixin):
             self.set_file_paths(self._file_paths)
 
             try:
-                self.log.debug("Removing old import errors")
-                DagFileProcessorManager.clear_nonexistent_import_errors(
-                    file_paths=self._file_paths, processor_subdir=self.get_dag_directory()
-                )
+                self.clear_nonexistent_import_errors()
             except Exception:
                 self.log.exception("Error removing old import errors")
 
@@ -702,24 +699,19 @@ class DagFileProcessorManager(LoggingMixin):
                 self._log_file_processing_stats(self._file_paths)
             self.last_stat_print_time = time.monotonic()
 
-    @staticmethod
     @provide_session
-    def clear_nonexistent_import_errors(
-        file_paths: list[str] | None, processor_subdir: str | None, session=NEW_SESSION
-    ):
+    def clear_nonexistent_import_errors(self, session=NEW_SESSION):
         """
         Clear import errors for files that no longer exist.
 
         :param file_paths: list of paths to DAG definition files
         :param session: session for ORM operations
         """
-        query = delete(ParseImportError)
+        self.log.debug("Removing old import errors")
+        query = delete(ParseImportError).where(ParseImportError.processor_subdir == self.get_dag_directory())
 
-        if file_paths:
-            query = query.where(
-                ~ParseImportError.filename.in_(file_paths),
-                ParseImportError.processor_subdir == processor_subdir,
-            )
+        if self._file_paths:
+            query = query.where(ParseImportError.filename.notin_(self._file_paths))
 
         session.execute(query.execution_options(synchronize_session="fetch"))
         session.commit()