+++ providers/src/airflow/providers/edge/executors/edge_executor.py
@@ -190,11 +190,18 @@ class EdgeExecutor(BaseExecutor):
                         TaskInstanceState.SUCCESS,
                         TaskInstanceState.FAILED,
                         TaskInstanceState.REMOVED,
+                        TaskInstanceState.RESTARTING,
+                        TaskInstanceState.UP_FOR_RETRY,
                     ]
                 )
             )
             .all()
         )
+
+        # Sync DB with executor otherwise runs out of sync in multi scheduler deployment
+        already_removed = self.running - set(job.key for job in jobs)
+        self.running = self.running - already_removed
+
         for job in jobs:
             if job.key in self.running:
                 if job.state == TaskInstanceState.RUNNING:
@@ -208,7 +215,11 @@ class EdgeExecutor(BaseExecutor):
                     if job.key in self.last_reported_state:
                         del self.last_reported_state[job.key]
                     self.success(job.key)
-                elif job.state == TaskInstanceState.FAILED:
+                elif job.state in [
+                    TaskInstanceState.FAILED,
+                    TaskInstanceState.RESTARTING,
+                    TaskInstanceState.UP_FOR_RETRY,
+                ]:
                     if job.key in self.last_reported_state:
                         del self.last_reported_state[job.key]
                     self.fail(job.key)
@@ -218,7 +229,13 @@ class EdgeExecutor(BaseExecutor):
                 job.state == TaskInstanceState.SUCCESS
                 and job.last_update_t < (datetime.now() - timedelta(minutes=job_success_purge)).timestamp()
             ) or (
-                job.state in (TaskInstanceState.FAILED, TaskInstanceState.REMOVED)
+                job.state
+                in (
+                    TaskInstanceState.FAILED,
+                    TaskInstanceState.REMOVED,
+                    TaskInstanceState.RESTARTING,
+                    TaskInstanceState.UP_FOR_RETRY,
+                )
                 and job.last_update_t < (datetime.now() - timedelta(minutes=job_fail_purge)).timestamp()
             ):
                 if job.key in self.last_reported_state: